+++ src/card/createCardElement.js
@@ -1,0 +1,10 @@
+import {
+  createElement
+} from '../utils.js';
+import {
+  createCardTemplate
+} from './createCardTemplate.js';
+
+
+export const createCardElement = (elementArray) =>
+  createElement(createCardTemplate(elementArray));
